telegram_game/requirements.txt
fastapi
uvicorn
python-telegram-bot
pygame
jinja2